+++ QuChu/app/src/main/java/co/quchu/quchu/model/PostCardDetailModel.java
@@ -50,7 +50,7 @@ public class PostCardDetailModel implements Serializable{
     private String plcaeName="";
     private int praiseNum;
     private String rgb;
-    private int score;
+    private float score;
     private String tel;
     private String time;
     /**
@@ -128,7 +128,7 @@ public class PostCardDetailModel implements Serializable{
         this.rgb = rgb;
     }
 
-    public void setScore(int score) {
+    public void setScore(float score) {
         this.score = score;
     }
 
@@ -208,7 +208,7 @@ public class PostCardDetailModel implements Serializable{
         return rgb;
     }
 
-    public int getScore() {
+    public float getScore() {
         return score;
     }
